Added send mail example
There are examples all over the internet as to how to send email with Java. This is another one, and there's nothing special about it. It's used to demonstrate how Documaker can generate emails that use the "preheader" which some email clients display. This isn't even needed to generate a preheader or do anything with it. It's just handy for sending email for SMTP servers that require authentication ;-)